+++ internal/compiler/generator/cpp.rs
@@ -3048,14 +3048,8 @@ fn compile_builtin_function_call(
             ctx.generator_state.conditional_includes.cstdlib.set(true);
             format!("[](const auto &a){{ auto e1 = std::end(a); auto e2 = const_cast<char*>(e1); auto r = std::strtod(std::begin(a), &e2); return e1 == e2 ? r : 0; }}({})", a.next().unwrap())
         }
-        BuiltinFunction::ColorRed => {
-            format!("{}.red()", a.next().unwrap())
-        }
-        BuiltinFunction::ColorGreen => {
-            format!("{}.green()", a.next().unwrap())
-        }
-        BuiltinFunction::ColorBlue => {
-            format!("{}.blue()", a.next().unwrap())
+        BuiltinFunction::ColorRGBAComponents => {
+            format!("{}", a.next().unwrap())
         }
         BuiltinFunction::ColorHue => {
             format!("{}.hue()", a.next().unwrap())