+++ CHANGELOG.md
@@ -23,7 +23,7 @@ All notable changes to this project are documented in this file.
  - Image: Added `horizontal-` and `vertical-tiling`
  - Flickable: Added `flicked` callback
  - Cupertino TabWidget: Tweaked visual appearance
- - Slint: Expose `.red()`, `.green()`, `.blue()` methods on `color`
+ - Slint: Expose `.red`, `.green`, `.blue` properties on `color`
  - Slint: Expose `.hue()`, `.saturation()`, `.brightness()` methods on `color`
  - Slint: Add `hsv` and `hsva` as method to create colors